labspace-exercises/tcc-basics/python/requirements.txt
# Core dependencies
testcontainers==3.7.1
psycopg2-binary==2.9.7
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
sqlalchemy==2.0.23

# Utilities
python-dotenv==1.0.0
dataclasses-json==0.6.1

# Development dependencies
black==23.7.0
flake8==6.0.0
mypy==1.5.1
